Log Redis cache stats in every app process
